[Unit] Attack Feedback

+++ DadVSMeClient/Assets/01.Scripts/Runtime/Core/GameSettings.cs
@@ -7,7 +7,7 @@ namespace DadVSMe
         private static float masterVolume = float.MinValue;
         public static float MasterVolume {
             get {
-                if(masterVolume == int.MinValue)
+                if(masterVolume == float.MinValue)
                     masterVolume = PlayerPrefs.GetFloat("MasterVolume", 1f);
 
                 return masterVolume;
@@ -22,7 +22,7 @@ namespace DadVSMe
         private static float bgmVolume = float.MinValue;
         public static float BGMVolume {
             get {
-                if(bgmVolume == int.MinValue)
+                if(bgmVolume == float.MinValue)
                     bgmVolume = PlayerPrefs.GetFloat("BGMVolume", 1f);
 
                 return bgmVolume;
@@ -37,7 +37,7 @@ namespace DadVSMe
         private static float sfxVolume = float.MinValue;
         public static float SFXVolume {
             get {
-                if(sfxVolume == int.MinValue)
+                if(sfxVolume == float.MinValue)
                     sfxVolume = PlayerPrefs.GetFloat("SFXVolume", 1f);
 
                 return sfxVolume;
